Moved the pause to the NUnitTestAssemblyRunner
